# Suggestions for performance improvements