--- a/main.go
+++ b/main.go
@@ -10,15 +10,6 @@ import (
 	"text/tabwriter"
 )
 
-func inc(ip net.IP) {
-	for j := len(ip) - 1; j >= 0; j-- {
-		ip[j]++
-		if ip[j] > 0 {
-			break
-		}
-	}
-}
-
 func main() {
 	log.SetPrefix("")
 	log.SetFlags(0)