Rather than incrementally testing each IP address in the network, you can directly calculate the answers in constant time. It results in a noticeable performance gain on especially large subnets like `/8`.

### Original
```
time ./ipcalc_original 10.0.0.0/8
CIDR Range: 10.0.0.0/8
Netmask:    255.0.0.0
First IP:   10.0.0.0
Last IP:    10.255.255.255
Addresses:  16777216

________________________________________________________
Executed in    2.09 secs    fish           external
   usr time    2.95 secs  123.00 micros    2.95 secs
   sys time    0.47 secs  769.00 micros    0.47 secs
```

### Suggested
```
time ./ipcalc_suggestions 10.0.0.0/8
CIDR Range: 10.0.0.0/8
Netmask:    255.0.0.0
First IP:   10.0.0.0
Last IP:    10.255.255.255
Addresses:  16777216

________________________________________________________
Executed in    6.49 millis    fish           external
   usr time    2.15 millis  112.00 micros    2.04 millis
   sys time    3.18 millis  667.00 micros    2.51 millis
```